MacGui: enable CLANG_WARN_QUOTED_INCLUDE_IN_FRAMEWORK_HEADER warning and fix the related warnings. Improved toolbars and alerts when compiling with the macOS 11 sdk.

+++ macosx/HBQueueMultiSelectionViewController.m
@@ -28,7 +28,7 @@
     }
     else
     {
-        self.label.stringValue = [NSString stringWithFormat:NSLocalizedString(@"%llu jobs selected", @""), self.count];
+        self.label.stringValue = [NSString stringWithFormat:NSLocalizedString(@"%lu jobs selected", @""), (unsigned long)self.count];
     }
 }
